requirements.txt
pillow
numpy
colorlog
pycryptodome
wxPython
pyperclip
natsort
orjson
stego-lsb==1.3.5
